(shitty "as it" commit: kibo and stylus
KIBO:

is it worth to create a lib to manage keyboard events? it seems to be so
let's test them.

* keypressed seems to be just overkill so i removed it
* kibo seems to be closer than what i expect from a simple abstracted
  keycode table

STYLUS:

CSS sucks as well as all the web stack so we need a preprocessor. i
spent some time reading some alternatives and it seems stylus is
the coolest syntax for what i expect.

THIS shitty commit:

this is just a WIP ... but we haven't touched pocamix for a while and i
need to show the state of art to members of the sxb.so